[otbnsim] Factor out test compile into shared lib
The otbnsim pytests typically need to assemble/compile source code for
testing. Factor this functionality out into a separate module to make it
easier to reuse in other test files.
