+++ hw/ip/otbn/dv/otbnsim/test/conftest.py
@@ -1,0 +1,9 @@
+# Copyright lowRISC contributors.
+# Licensed under the Apache License, Version 2.0, see LICENSE for details.
+# SPDX-License-Identifier: Apache-2.0
+
+import os
+import sys
+
+# Make tested code available for import
+sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
